requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
psycopg2-binary==2.9.7
requests==2.31.0
yookassa==3.0.1
vk_api==11.9.9
